requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psutil==7.1.3
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
pandas
openpyxl